requirements-dev.txt
pycodestyle==2.5.0
pytest==5.3.5
pytest-django==3.8.0
selenium==3.141.0
django-selenium-test==1.0.2